requirements.txt
datasets==2.21.0
google-generativeai==0.7.2
ipykernel==6.29.5
ipython==8.26.0
jsonlines==4.0.0
meilisearch==0.31.5
openai==1.42.0
python-dotenv==1.0.1
streamlit==1.37.1